Nest the recursion in Control differently.
This change facilitates proofs with stmt_eval. The previous scheme gave
hypotheses of the form
    stmt_eval menv env (Control mems c (Ifte b s Skip)) (menv', env')

whereas the new scheme gives ones of the form
    stmt_eval menv env (Ifte b (Control mems c s) Skip) (menv', env')

which can be inverted directly with Iifte_true/Iifte_false, making
inductions much easier.

The resulting code has the same effect, although the conditionals are nested
in the reverse order. There remains the question of which order is
preferable (for collapsing conditionals to minimize them).