Refine Is_well_sch for node applications.
A variable (x) defined by an application equation (EqApp x _ _), must not
have already been defined earlier by an application equation. This
duplication is unproblematic for EqDef and EqFby since the existence of a
semantics ensures a consistent assignment. The problem with EqApp is that
the translated imperative code rewrites the associated instance memory,
doing so twice in a row is (in general) not allowed!